+++ configure.sh
@@ -24,6 +24,7 @@ for_debian() {
 	check_installed libx11-dev /usr/include/X11/Xlib.h libx11-dev
 	check_installed libsqlite3-dev /usr/include/sqlite3.h libsqlite3-dev
 	check_installed libgdk-pixbuf2.0-dev /usr/include/gdk-pixbuf-2.0/gdk-pixbuf/gdk-pixbuf.h libgdk-pixbuf2.0-dev
+	check_installed libasound2-dev /usr/include/alsa/asoundlib.h libasound2-dev
 	check_installed xterm /usr/bin/xterm xterm
 
 	if [ "$packages" != "" ]
@@ -43,6 +44,7 @@ for_arch() {
 	check_installed libx11-dev /usr/include/X11/Xlib.h libx11-dev
 	check_installed libsqlite3-dev /usr/include/sqlite3.h libsqlite3-dev
 	check_installed gdk-pixbuf2 /usr/include/gdk-pixbuf-2.0/gdk-pixbuf/gdk-pixbuf.h gdk-pixbuf2
+	check_installed alsa-lib /usr/include/alsa/asoundlib.h alsa-lib
 	check_installed xterm /usr/bin/xterm xterm
 
 	if [ "$packages" != "" ]