+++ sound.h
@@ -1,0 +1,8 @@
+#ifndef SOUND_H
+#define SOUND_H
+void sound_init(void);
+int sound_get_volume(void);
+void sound_volume_up(void);
+void sound_volume_down(void);
+void sound_show(void);
+#endif